исправление косяков и приведение к крутому виду

+++ include/core/model/MeterList.h
@@ -1,16 +1,18 @@
-// include/meters/MeterList.h
 #pragma once
+
 #include <vector>
 #include <memory>
 #include "core/model/Meters.h"
+
 
 class MeterList {
 public:
     void addMeter(std::unique_ptr<AbstractMeter> meter);
     void removeMeter(size_t index);
     void clear();
+
     const std::vector<std::unique_ptr<AbstractMeter>>& getMeters() const;
-    size_t size() const;
+    int size() const;
     bool empty() const;
 
 private:
